Working on moving code from op mode to BotFactory and making BotConfig and it's subclass of summerBot
