+++ TeamCode/src/main/java/edu/ahs/robotics/Chassis.java
@@ -1,13 +1,11 @@
 package edu.ahs.robotics;
 
-import org.firstinspires.ftc.robotcore.internal.android.dx.util.Warning;
-
 /**
  * The Chassis class is the superclass for all chassis types
  */
 public abstract class Chassis implements Executor{
 
-    public Chassis(BotConfig botConfig){
+    public Chassis(BotFactory botFactory){
     }
     public abstract void execute(PlanElement planElement);
