+++ TeamCode/src/main/java/edu/ahs/robotics/MecanumChassis.java
@@ -23,16 +23,16 @@ public class MecanumChassis extends Chassis {
         }
     }
 
-    public MecanumChassis(BotConfig botConfig) {
-        super(botConfig);
+    public MecanumChassis(BotFactory botFactory) {
+        super(botFactory);
 
-        HashMap<MotorLocations, String> deviceNames = botConfig.getDriveMotorDeviceNames();
-        //HashMap<MotorLocations, Boolean> deviceFlips = botConfig.getFlippedMotors();
+        HashMap<MotorLocations, String> deviceNames = botFactory.getDriveMotorDeviceNames();
+        //HashMap<MotorLocations, Boolean> deviceFlips = botFactory.getFlippedMotors();
 
-        frontLeft = new SingleDriveUnit(deviceNames.get(MotorLocations.FRONTLEFT), botConfig, botConfig.isFlipped(MotorLocations.FRONTLEFT));
-        frontRight = new SingleDriveUnit(deviceNames.get(MotorLocations.FRONTRIGHT), botConfig, botConfig.isFlipped(MotorLocations.FRONTRIGHT));
-        backLeft = new SingleDriveUnit(deviceNames.get(MotorLocations.BACKLEFT), botConfig, botConfig.isFlipped(MotorLocations.BACKLEFT));
-        backRight = new SingleDriveUnit(deviceNames.get(MotorLocations.BACKRIGHT), botConfig, botConfig.isFlipped(MotorLocations.BACKRIGHT));
+        frontLeft = new SingleDriveUnit(deviceNames.get(MotorLocations.FRONTLEFT), botFactory, botFactory.isFlipped(MotorLocations.FRONTLEFT));
+        frontRight = new SingleDriveUnit(deviceNames.get(MotorLocations.FRONTRIGHT), botFactory, botFactory.isFlipped(MotorLocations.FRONTRIGHT));
+        backLeft = new SingleDriveUnit(deviceNames.get(MotorLocations.BACKLEFT), botFactory, botFactory.isFlipped(MotorLocations.BACKLEFT));
+        backRight = new SingleDriveUnit(deviceNames.get(MotorLocations.BACKRIGHT), botFactory, botFactory.isFlipped(MotorLocations.BACKRIGHT));
     }
     public void turnOnMotor(){
         frontLeft.setPower(1);
